Implement overbounce mutator
Unfortunately it's a bit of a fake overbounce because it has to create a
condition for overbouncing. The condition is modeled after the *real*
Q3 overbounce condition, and it is sufficiently random yet
reproducible so map makers can create areas where OB is possible.

The movement code responsible for it is ported straight from Q3 and
so the mechanics remain mostly the same. Vertical and horizontal OB
is possible, though horizontal is a bit broken due to how strong
Minecraft gravity is.